Add files via upload
Robot Code in Arduino using the following:
1. An arduino Mega controller.
2. Ubloxx GPS module
3. A ubloxx compass module
4. An arduino GSM shield

The code takes as input destination or intermediate coordinates that it saves as waypoints. It uses the haversine heading and distance formulaes to determine the direction it needs to move into and the distance form destination. The robot code can be extended with obstacle detection to allow our robot to maneuver on roads.