Add files via upload
This is a code based on the A* route navigation algorithm used to calculate intermediate GPS coordinates given initial and final coordinates. The code uses an OSM file to look up GPS coordinates which can be downloaded from OpenstreetMaps for any map segment you want. Its opensource so allows you to create your own map as was done for our project.