Add files via upload
This is the main controller Code. This allows you send via text messages the GPS coordinates for your final location. You can do that by using an Android phone application. Also it uses an Arduinio Mega controller and shield integrated with a CPU via USB. The controller also keeps track of any robots that are linked to it. It sends the GPS destination and robot coordinate serially to the computer its connected to help calculate the intermediate path coordinates for the robot.